# KT-51492 fix benchmark annotations