Before:
```
Benchmark                     (size)  Mode  Cnt       Score       Error  Units
LoopBenchmark.arrayIndexLoop      10  avgt    5      74,313 ?     0,496  ns/op
LoopBenchmark.arrayIndexLoop    1000  avgt    5    6947,446 ?    60,806  ns/op
LoopBenchmark.arrayIndexLoop  100000  avgt    5  694597,949 ? 10135,009  ns/op
```

After:
```
Benchmark                     (size)  Mode  Cnt       Score      Error  Units
LoopBenchmark.arrayIndexLoop      10  avgt    5      41,933 ?    0,425  ns/op
LoopBenchmark.arrayIndexLoop    1000  avgt    5    3851,127 ?   51,661  ns/op
LoopBenchmark.arrayIndexLoop  100000  avgt    5  408370,300 ? 2858,830  ns/op
```